trainer/requirements.txt
aim==3.17.5
aim-ui==3.17.5
aimrecords==0.0.7
aimrocks==0.4.0
aiofiles==23.2.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
backoff==2.2.1
base58==2.0.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.5
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.13.1
greenlet==3.0.3
grpcio==1.62.0
h11==0.14.0
idna==3.6
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
numpy==1.26.4
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
py3nvml==0.2.7
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
RestrictedPython==7.0
segment-analytics-python==2.3.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.52
starlette==0.36.3
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
xmltodict==0.13.0
